Fix /gtceu share_prospection_data command (#4919)
